#3031: Adding one test
